+++ neo/d3xp/Moveable.cpp
@@ -1146,7 +1146,7 @@ void idExplodingBarrel::Killed( idEntity *inflictor, idEntity *attacker, int dam
 			byte		msgBuf[MAX_EVENT_PARAM_SIZE];
 
 			msg.InitWrite( msgBuf, sizeof( msgBuf ) );
-			msg.WriteLong( gameLocal.time );
+			msg.WriteInt( gameLocal.time );
 			ServerSendEvent( EVENT_EXPLODE, &msg, false );
 		}
 	}
@@ -1343,7 +1343,7 @@ bool idExplodingBarrel::ClientReceiveEvent( int event, int time, const idBitMsg 
 
 	switch( event ) {
 		case EVENT_EXPLODE: {
-			if ( gameLocal.realClientTime - msg.ReadLong() < spawnArgs.GetInt( "explode_lapse", "1000" ) ) {
+			if ( gameLocal.realClientTime - msg.ReadInt() < spawnArgs.GetInt( "explode_lapse", "1000" ) ) {
 				ExplodingEffects( );
 			}
 			return true;
