+++ neo/idlib/math/Math.h
@@ -1406,11 +1406,11 @@ reciprocal square root, returns huge number when x == 0.0
 */
 ID_INLINE float idMath::RSqrt( float x ) {
 
-	long i;
+	int i;
 	float y, r;
 
 	y = x * 0.5f;
-	i = *reinterpret_cast<long *>( &x );
+	i = *reinterpret_cast<int *>( &x );
 	i = 0x5f3759df - ( i >> 1 );
 	r = *reinterpret_cast<float *>( &i );
 	r = r * ( 1.5f - r * r * y );
