+++ neo/idlib/math/Simd_SSE.cpp
@@ -273,13 +273,13 @@ void VPCALL idSIMD_SSE::BlendJoints( idJointQuat *joints, const idJointQuat *ble
 		float omega;
 		float scale0;
 		float scale1;
-		unsigned long signBit;
+		unsigned int signBit;
 
 		cosom = jointQuat.x * blendQuat.x + jointQuat.y * blendQuat.y + jointQuat.z * blendQuat.z + jointQuat.w * blendQuat.w;
 
-		signBit = (*(unsigned long *)&cosom) & ( 1 << 31 );
+		signBit = (*(unsigned int *)&cosom) & ( 1 << 31 );
 
-		(*(unsigned long *)&cosom) ^= signBit;
+		(*(unsigned int *)&cosom) ^= signBit;
 
 		scale0 = 1.0f - cosom * cosom;
 		scale0 = ( scale0 <= 0.0f ) ? 1e-10f : scale0;
@@ -288,7 +288,7 @@ void VPCALL idSIMD_SSE::BlendJoints( idJointQuat *joints, const idJointQuat *ble
 		scale0 = idMath::Sin16( ( 1.0f - lerp ) * omega ) * sinom;
 		scale1 = idMath::Sin16( lerp * omega ) * sinom;
 
-		(*(unsigned long *)&scale1) ^= signBit;
+		(*(unsigned int *)&scale1) ^= signBit;
 
 		jointQuat.x = scale0 * jointQuat.x + scale1 * blendQuat.x;
 		jointQuat.y = scale0 * jointQuat.y + scale1 * blendQuat.y;
@@ -487,7 +487,7 @@ idSIMD_SSE::ConvertJointQuatsToJointMats
 void VPCALL idSIMD_SSE::ConvertJointQuatsToJointMats( idJointMat *jointMats, const idJointQuat *jointQuats, const int numJoints ) {
 	assert( sizeof( idJointQuat ) == JOINTQUAT_SIZE );
 	assert( sizeof( idJointMat ) == JOINTMAT_SIZE );
-	assert( (int)(&((idJointQuat *)0)->t) == (int)(&((idJointQuat *)0)->q) + (int)sizeof( ((idJointQuat *)0)->q ) );
+	assert( (intptr_t)(&((idJointQuat *)0)->t) == (intptr_t)(&((idJointQuat *)0)->q) + (intptr_t)sizeof( ((idJointQuat *)0)->q ) );
 
 	const float * jointQuatPtr = (float *)jointQuats;
 	float * jointMatPtr = (float *)jointMats;
@@ -620,7 +620,7 @@ void VPCALL idSIMD_SSE::ConvertJointMatsToJointQuats( idJointQuat *jointQuats, c
 
 	assert( sizeof( idJointQuat ) == JOINTQUAT_SIZE );
 	assert( sizeof( idJointMat ) == JOINTMAT_SIZE );
-	assert( (int)(&((idJointQuat *)0)->t) == (int)(&((idJointQuat *)0)->q) + (int)sizeof( ((idJointQuat *)0)->q ) );
+	assert( (intptr_t)(&((idJointQuat *)0)->t) == (intptr_t)(&((idJointQuat *)0)->q) + (intptr_t)sizeof( ((idJointQuat *)0)->q ) );
 
 	const __m128 vector_float_zero		= _mm_setzero_ps();
 	const __m128 vector_float_one		= { 1.0f, 1.0f, 1.0f, 1.0f };
